# specs/001-multitenant-foundation/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Adsup Multi-tenant Foundation API
  version: 1.0.0
  description: Stable Module 1 contract. All tenant paths re-authorize active membership and scope.
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/tenants/{tenantId}/branches:
    get:
      tags: [Organization]
      operationId: listBranches
      x-permission: tenant.read
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      responses:
        '200':
          description: Tenant branches
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationUnitList' }
    post:
      tags: [Organization]
      operationId: createBranch
      x-permission: branch.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateBranch' }
      responses:
        '201':
          description: Branch created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationUnit' }
        '403': { $ref: '#/components/responses/ForbiddenProblem' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/tenants/{tenantId}/branches/{branchId}:
    patch:
      tags: [Organization]
      operationId: updateBranchStatus
      x-permission: branch.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: branchId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrganizationUnitStatus' }
      responses:
        '200':
          description: Branch lifecycle state updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationUnit' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/tenants/{tenantId}/departments:
    get:
      tags: [Organization]
      operationId: listDepartments
      x-permission: tenant.read
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      responses:
        '200':
          description: Tenant departments
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationUnitList' }
    post:
      tags: [Organization]
      operationId: createDepartment
      x-permission: organization.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganizationUnit' }
      responses:
        '201':
          description: Department created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationUnit' }
        '403': { $ref: '#/components/responses/ForbiddenProblem' }
  /v1/tenants/{tenantId}/departments/{departmentId}:
    patch:
      tags: [Organization]
      operationId: updateDepartmentStatus
      x-permission: organization.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: departmentId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrganizationUnitStatus' }
      responses:
        '200':
          description: Department lifecycle state updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationUnit' }
  /v1/tenants/{tenantId}/positions:
    get:
      tags: [Organization]
      operationId: listPositions
      x-permission: tenant.read
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      responses:
        '200':
          description: Tenant positions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationUnitList' }
    post:
      tags: [Organization]
      operationId: createPosition
      x-permission: organization.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateOrganizationUnit' }
      responses:
        '201':
          description: Position created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationUnit' }
  /v1/tenants/{tenantId}/positions/{positionId}:
    patch:
      tags: [Organization]
      operationId: updatePositionStatus
      x-permission: organization.manage
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - name: positionId
          in: path
          required: true
          schema: { $ref: '#/components/schemas/Uuid' }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateOrganizationUnitStatus' }
      responses:
        '200':
          description: Position lifecycle state updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrganizationUnit' }
...
  /v1/tenants/{tenantId}/media/upload-intents:
    post:
      tags: [Media]
      operationId: createMediaUploadIntent
      x-permission: media.create
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMediaIntent' }
      responses:
        '201':
          description: Pending media and short-lived signed upload
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaUploadIntent' }
        '403': { $ref: '#/components/responses/ForbiddenProblem' }
  /v1/tenants/{tenantId}/media/{mediaId}/complete:
    post:
      tags: [Media]
      operationId: completeMediaUpload
      x-permission: media.create
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/MediaId'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Verified ready media
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MediaObject' }
        '409': { $ref: '#/components/responses/ConflictProblem' }
  /v1/tenants/{tenantId}/media/{mediaId}/download-url:
    post:
      tags: [Media]
      operationId: createMediaDownloadUrl
      x-permission: media.read
      parameters:
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/MediaId'
      responses:
        '200':
          description: Short-lived signed download URL
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SignedUrl' }
        '404': { $ref: '#/components/responses/NotFoundProblem' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    MediaId:
      name: mediaId
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Uuid' }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 128 }
  responses:
...
    ForbiddenProblem:
      description: Actor lacks permission/scope
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFoundProblem:
      description: Resource absent or intentionally hidden outside scope
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ConflictProblem:
      description: State, concurrency, last-owner or idempotency fingerprint conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Uuid:
      type: string
      format: uuid
    Timestamp:
      type: string
      format: date-time
...
    OrganizationUnit:
      type: object
      required: [id, tenantId, code, name, status]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        tenantId: { $ref: '#/components/schemas/Uuid' }
        code: { type: string }
        name: { type: string }
        status: { enum: [ACTIVE, INACTIVE] }
    OrganizationUnitList:
      type: array
      items: { $ref: '#/components/schemas/OrganizationUnit' }
    CreateOrganizationUnit:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_-]{2,32}$' }
        name: { type: string, minLength: 2, maxLength: 160 }
    CreateBranch:
      type: object
      additionalProperties: false
      required: [code, name]
      properties:
        code: { type: string, pattern: '^[A-Z0-9_-]{2,32}$' }
        name: { type: string, minLength: 2, maxLength: 160 }
        timezoneOverride: { type: string, minLength: 3, maxLength: 64 }
    UpdateOrganizationUnitStatus:
      type: object
      additionalProperties: false
      required: [status, reason]
      properties:
        status: { enum: [ACTIVE, INACTIVE] }
        reason: { type: string, minLength: 3, maxLength: 500 }
...
    CreateMediaIntent:
      type: object
      additionalProperties: false
      required: [purpose, contentType, byteSize, checksumSha256]
      properties:
        purpose: { type: string, minLength: 2, maxLength: 50 }
        branchId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
        sourceType: { type: string, minLength: 2, maxLength: 80 }
        sourceId: { anyOf: [{$ref: '#/components/schemas/Uuid'}, {type: 'null'}] }
        contentType: { enum: [image/jpeg, image/png, image/webp, video/mp4, application/pdf] }
        byteSize: { type: integer, minimum: 1, maximum: 524288000 }
        checksumSha256: { type: string, pattern: '^[A-Fa-f0-9]{64}$' }
    MediaUploadIntent:
      type: object
      required: [media, upload]
      properties:
        media: { $ref: '#/components/schemas/MediaObject' }
        upload:
          $ref: '#/components/schemas/SignedUrl'
    MediaObject:
      type: object
      required: [id, tenantId, purpose, contentType, byteSize, checksumSha256, status]
      properties:
        id: { $ref: '#/components/schemas/Uuid' }
        tenantId: { $ref: '#/components/schemas/Uuid' }
        purpose: { type: string }
        contentType: { type: string }
        byteSize: { type: integer }
        checksumSha256: { type: string }
        status: { enum: [PENDING_UPLOAD, READY, REJECTED, DELETED] }
    SignedUrl:
      type: object
      required: [url, expiresAt, requiredHeaders]
      properties:
        url: { type: string, format: uri }
        expiresAt: { $ref: '#/components/schemas/Timestamp' }
        requiredHeaders:
          type: object
          additionalProperties: { type: string }
...
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, correlationId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: '^[A-Z0-9_]+$' }
        detail: { type: string }
        correlationId: { type: string }
        fieldErrors:
          type: array
          items:
            type: object
            required: [path, code, message]
            properties:
              path: { type: string }
              code: { type: string }
              message: { type: string }
security:
  - bearerAuth: []